finalizaing cuda stream, separating state C and C++ interfaces
