+++ cmake/dependencies.cmake
@@ -79,6 +79,7 @@ if (MUSICA_ENABLE_MICM AND MUSICA_BUILD_C_CXX_INTERFACE)
   set(MICM_ENABLE_EXAMPLES OFF)
   set(MICM_DEFAULT_VECTOR_MATRIX_SIZE ${MUSICA_SET_MICM_VECTOR_MATRIX_SIZE})
   set(MICM_ENABLE_CUDA ${MUSICA_ENABLE_CUDA})
+  set(MICM_GPU_TYPE a100)
 
   FetchContent_MakeAvailable(micm)
 endif()
